feat: 0.2.0 reliability, reflash, and build footer
Infer hardware running from UART time (physical start/stop), clamp
scores 0-99, LRU/NAK DHCP, strip EDNS from DNS replies, continue after
panic-persist, and feed the RP2040 watchdog. Add just reflash via USB
CDC ENTERBOOTLOADER, git-testament version/hash/date in the UI, and
rename the binary to scoreboard-ctrl.

+++ src/ap_log.rs
@@ -69,7 +69,7 @@ pub fn format_qname(pkt: &[u8], out: &mut [u8]) -> usize {
             if i + 1 >= pkt.len() {
                 break;
             }
-            i = (((len & 0x3F) as usize) << 8) | (pkt[i + 1] as usize);
+            i = ((len & 0x3F) << 8) | (pkt[i + 1] as usize);
             continue;
         }
         if i + 1 + len > pkt.len() {
